Integration of the tus based upload, move to php 7 and nginx
- Changed the base docker image to php7 fpm and moved from Apache to Nginx for the internal webserver
- Tus based upload is accessible via the create or add menu in the documents section
- dropped /dms requirement
- Added supervisor to monitor the startup and the shutdown of the services inside the K-Box docker image
- redirect /dms routes to the root for backward compatibility after /dms being deprecated

+++ app/Http/Controllers/Document/DocumentsController.php
@@ -840,22 +840,24 @@ class DocumentsController extends Controller
                 if ($document->isDirty() || $group_dirty) {
                     $document->save();
 
-                    $this->service->reindexDocument($document, \KlinkVisibilityType::KLINK_PRIVATE);
-                    
-                    if ($user->can_capability(Capability::CHANGE_DOCUMENT_VISIBILITY)) {
-                        if (! $was_document_public && $document->is_public) {
-                            \Log::info('Applying visibility change', ['descriptor' => $document->id, 'old' => $was_document_public, 'new' => $document->is_public]);
-                            $this->service->reindexDocument($document, \KlinkVisibilityType::KLINK_PUBLIC);
-                        } elseif ($was_document_public && ! $document->is_public) {
-                            \Log::info('Applying visibility change', ['descriptor' => $document->id, 'old' => $was_document_public, 'new' => $document->is_public]);
-                            $this->service->deletePublicDocument($document);
+                    if ($document->isFileUploadComplete()) {
+                        $this->service->reindexDocument($document, \KlinkVisibilityType::KLINK_PRIVATE);
+                        
+                        if ($user->can_capability(Capability::CHANGE_DOCUMENT_VISIBILITY)) {
+                            if (! $was_document_public && $document->is_public) {
+                                \Log::info('Applying visibility change', ['descriptor' => $document->id, 'old' => $was_document_public, 'new' => $document->is_public]);
+                                $this->service->reindexDocument($document, \KlinkVisibilityType::KLINK_PUBLIC);
+                            } elseif ($was_document_public && ! $document->is_public) {
+                                \Log::info('Applying visibility change', ['descriptor' => $document->id, 'old' => $was_document_public, 'new' => $document->is_public]);
+                                $this->service->deletePublicDocument($document);
+                            } elseif ($was_document_public && $document->is_public) {
+                                \Log::info('Reindexing also Public Document because Doc is dirty', ['descriptor' => $document->id]);
+                                $this->service->reindexDocument($document, \KlinkVisibilityType::KLINK_PUBLIC);
+                            }
                         } elseif ($was_document_public && $document->is_public) {
                             \Log::info('Reindexing also Public Document because Doc is dirty', ['descriptor' => $document->id]);
                             $this->service->reindexDocument($document, \KlinkVisibilityType::KLINK_PUBLIC);
                         }
-                    } elseif ($was_document_public && $document->is_public) {
-                        \Log::info('Reindexing also Public Document because Doc is dirty', ['descriptor' => $document->id]);
-                        $this->service->reindexDocument($document, \KlinkVisibilityType::KLINK_PUBLIC);
                     }
                 } else {
                     $document->touch();
